requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
Click==7.0
Deprecated==1.2.7
Flask==1.1.1
google==2.0.3
httplib2==0.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
kwconfig==1.1.7
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.45
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dotenv==0.19.2
pytz==2019.3
rauth==0.7.3
requests==2.22.0
rsa==4.7
six==1.13.0
soupsieve==1.9.5
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
